relocate descriptor
