+++ internal/descriptor/cache/cache.go
@@ -2,8 +2,8 @@ package cache
 
 import (
 	"sync"
-
-	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
+	
+	_type "github.com/kyma-project/lifecycle-manager/internal/descriptor/type"
 )
 
 type DescriptorCache struct {
@@ -16,19 +16,19 @@ func NewDescriptorCache() *DescriptorCache {
 	}
 }
 
-func (d *DescriptorCache) Get(key DescriptorKey) *v1beta2.Descriptor {
+func (d *DescriptorCache) Get(key DescriptorKey) *_type.Descriptor {
 	value, ok := d.cache.Load(key)
 	if !ok {
 		return nil
 	}
-	desc, ok := value.(*v1beta2.Descriptor)
+	desc, ok := value.(*_type.Descriptor)
 	if !ok {
 		return nil
 	}
 
-	return &v1beta2.Descriptor{ComponentDescriptor: desc.Copy()}
+	return &_type.Descriptor{ComponentDescriptor: desc.Copy()}
 }
 
-func (d *DescriptorCache) Set(key DescriptorKey, value *v1beta2.Descriptor) {
+func (d *DescriptorCache) Set(key DescriptorKey, value *_type.Descriptor) {
 	d.cache.Store(key, value)
 }
