+++ internal/descriptor/cache/cache_test.go
@@ -7,8 +7,8 @@ import (
 	"ocm.software/ocm/api/ocm/compdesc"
 	ocmmetav1 "ocm.software/ocm/api/ocm/compdesc/meta/v1"
 
-	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
 	"github.com/kyma-project/lifecycle-manager/internal/descriptor/cache"
+	"github.com/kyma-project/lifecycle-manager/internal/descriptor/type"
 )
 
 func TestGet_ForCacheWithoutEntry_ReturnsNoEntry(t *testing.T) {
@@ -30,7 +30,7 @@ func TestGet_ForCacheWithAnEntry_ReturnsAnEntry(t *testing.T) {
 			},
 		},
 	}
-	desc1 := &v1beta2.Descriptor{ComponentDescriptor: ocmDesc1}
+	desc1 := &_type.Descriptor{ComponentDescriptor: ocmDesc1}
 
 	descriptorCache.Set(key1, desc1)
 
@@ -39,14 +39,14 @@ func TestGet_ForCacheWithAnEntry_ReturnsAnEntry(t *testing.T) {
 
 func TestGet_ForCacheWithOverwrittenEntry_ReturnsNewEntry(t *testing.T) {
 	descriptorCache := cache.NewDescriptorCache()
-	originalKey, originalValue := cache.DescriptorKey("key 1"), &v1beta2.Descriptor{
+	originalKey, originalValue := cache.DescriptorKey("key 1"), &_type.Descriptor{
 		ComponentDescriptor: &compdesc.ComponentDescriptor{
 			ComponentSpec: compdesc.ComponentSpec{
 				ObjectMeta: ocmmetav1.ObjectMeta{Name: "descriptor 1"},
 			},
 		},
 	}
-	newKey, newValue := cache.DescriptorKey("key 2"), &v1beta2.Descriptor{
+	newKey, newValue := cache.DescriptorKey("key 2"), &_type.Descriptor{
 		ComponentDescriptor: &compdesc.ComponentDescriptor{
 			ComponentSpec: compdesc.ComponentSpec{
 				ObjectMeta: ocmmetav1.ObjectMeta{Name: "descriptor 2"},
@@ -64,14 +64,14 @@ func TestGet_ForCacheWithOverwrittenEntry_ReturnsNewEntry(t *testing.T) {
 	assertDescriptorEqual(t, newValue, descriptorCache.Get(originalKey))
 }
 
-func assertDescriptorEqual(t *testing.T, expected, actual *v1beta2.Descriptor) {
+func assertDescriptorEqual(t *testing.T, expected, actual *_type.Descriptor) {
 	t.Helper()
 	if expected.Name != actual.Name {
 		t.Fatalf("Expected and actual descriptors do not match: \nExpected: %#v \nActual: %#v", expected, actual)
 	}
 }
 
-func assertDescriptorNotEqual(t *testing.T, expected, actual *v1beta2.Descriptor) {
+func assertDescriptorNotEqual(t *testing.T, expected, actual *_type.Descriptor) {
 	t.Helper()
 	if expected.Name == actual.Name {
 		t.Fatalf("Expected and actual descriptors do match: \nExpected: %#v \nActual: %#v", expected, actual)
