+++ internal/descriptor/provider/provider.go
@@ -2,6 +2,7 @@ package provider
 
 import (
 	"errors"
+	"github.com/kyma-project/lifecycle-manager/internal/descriptor/type"
 
 	"ocm.software/ocm/api/ocm/compdesc"
 
@@ -26,13 +27,13 @@ func NewCachedDescriptorProvider() *CachedDescriptorProvider {
 	}
 }
 
-func (c *CachedDescriptorProvider) GetDescriptor(template *v1beta2.ModuleTemplate) (*v1beta2.Descriptor, error) {
+func (c *CachedDescriptorProvider) GetDescriptor(template *v1beta2.ModuleTemplate) (*_type.Descriptor, error) {
 	if template == nil {
 		return nil, ErrTemplateNil
 	}
 
 	if template.Spec.Descriptor.Object != nil {
-		desc, ok := template.Spec.Descriptor.Object.(*v1beta2.Descriptor)
+		desc, ok := template.Spec.Descriptor.Object.(*_type.Descriptor)
 		if !ok {
 			return nil, ErrTypeAssert
 		}
@@ -56,8 +57,8 @@ func (c *CachedDescriptorProvider) GetDescriptor(template *v1beta2.ModuleTemplat
 		return nil, errors.Join(ErrDecode, err)
 	}
 
-	template.Spec.Descriptor.Object = &v1beta2.Descriptor{ComponentDescriptor: ocmDesc}
-	descriptor, ok := template.Spec.Descriptor.Object.(*v1beta2.Descriptor)
+	template.Spec.Descriptor.Object = &_type.Descriptor{ComponentDescriptor: ocmDesc}
+	descriptor, ok := template.Spec.Descriptor.Object.(*_type.Descriptor)
 	if !ok {
 		return nil, ErrTypeAssert
 	}
@@ -76,7 +77,7 @@ func (c *CachedDescriptorProvider) Add(template *v1beta2.ModuleTemplate) error {
 	}
 
 	if template.Spec.Descriptor.Object != nil {
-		desc, ok := template.Spec.Descriptor.Object.(*v1beta2.Descriptor)
+		desc, ok := template.Spec.Descriptor.Object.(*_type.Descriptor)
 		if ok && desc != nil {
 			c.DescriptorCache.Set(key, desc)
 			return nil
@@ -90,8 +91,8 @@ func (c *CachedDescriptorProvider) Add(template *v1beta2.ModuleTemplate) error {
 		return errors.Join(ErrDecode, err)
 	}
 
-	template.Spec.Descriptor.Object = &v1beta2.Descriptor{ComponentDescriptor: ocmDesc}
-	descriptor, ok := template.Spec.Descriptor.Object.(*v1beta2.Descriptor)
+	template.Spec.Descriptor.Object = &_type.Descriptor{ComponentDescriptor: ocmDesc}
+	descriptor, ok := template.Spec.Descriptor.Object.(*_type.Descriptor)
 	if !ok {
 		return ErrTypeAssert
 	}
