+++ internal/descriptor/type/descriptor.go
@@ -1,0 +1,31 @@
+package _type
+
+import (
+	machineryruntime "k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"ocm.software/ocm/api/ocm/compdesc"
+)
+
+type Descriptor struct {
+	*compdesc.ComponentDescriptor
+}
+
+func (d *Descriptor) SetGroupVersionKind(kind schema.GroupVersionKind) {
+	d.Version = kind.Version
+}
+
+func (d *Descriptor) GroupVersionKind() schema.GroupVersionKind {
+	return schema.GroupVersionKind{
+		Group:   "ocm.kyma-project.io",
+		Version: d.Metadata.ConfiguredVersion,
+		Kind:    "Descriptor",
+	}
+}
+
+func (d *Descriptor) GetObjectKind() schema.ObjectKind {
+	return d
+}
+
+func (d *Descriptor) DeepCopyObject() machineryruntime.Object {
+	return &Descriptor{ComponentDescriptor: d.Copy()}
+}
